// src/store/modules/type.ts
export enum Types {
  TRIGGER_CREATE = 'TRIGGER_CREATE',
  TRIGGER_SAVE = 'TRIGGER_SAVE',
  TRIGGER_PUBLISH = 'TRIGGER_PUBLISH',
  SET_PROCESS_DETAIL = 'SET_PROCESS_DETAIL',
  SHOW_GLOBAL = 'SHOW_GLOBAL',
  SHOW_NODE = 'SHOW_NODE',
  SHOW_REL = 'SHOW_REL',
  UPDATE_NODE_DATA = 'UPDATE_NODE_DATA',
}